Remove guarantee that a data entry exists within remote messages
Currently, the `data` property within the RemoteMessage type can lead to runtime errors if the user does not manually check that the value exists as a string. This change adds ambiguity to prevent users from mistreating a value as a string, when it's not guaranteed to exist.

Example of an error:
Assume a remote message only contains this data: `{firstName, lastName}`
If the user accesses a non-existant property, e.g. `data.fName`, the compiler would infer that `fName` is a string.
Calling any string method, such as `toLowerCase()`, on that value would cause a runtime error.